refactor: Refactor into services
