+++ coffeehr/coffeehr_service.py
@@ -4,9 +4,12 @@ from datetime import datetime, date
 from compose import compose
 import pytz
 
+from db import bigquery
 from coffeehr.pipeline.interface import Pipeline, Data
-from coffeehr.repo import get
-from db.bigquery import load
+from coffeehr.pipeline import pipelines
+from coffeehr.coffeehr_repo import get
+from tasks import cloud_tasks
+
 
 TZ = pytz.timezone("Asia/Ho_Chi_Minh")
 
@@ -41,8 +44,19 @@ def pipeline_service(pipeline: Pipeline) -> int:
     start = date(2017, 1, 1)
     end = date(date.today().year + 1, 1, 1)
     return compose(
-        load(pipeline.name, pipeline.schema),
+        bigquery.load(pipeline.name, pipeline.schema),
         _transform_timestamp(pipeline.schema),
         pipeline.transform,
         get(pipeline.get_options),
     )((start, end))
+
+
+
+def create_tasks_service() -> dict[str, int]:
+    return {
+        "tasks": cloud_tasks.create_tasks(
+            "coffeehr",
+            [{"table": table} for table in pipelines.keys()],
+            lambda x: x["table"],
+        )
+    }
