+++ pyproject.toml
@@ -25,9 +25,6 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
 addopts = "-s --durations=0"
-testpaths = [
-    "test",
-]
 
 [tool.mypy]
 python_version = 3.9
